pyproject.toml
[project]
name = "identify-doc-gaps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.10.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "blis==1.3.0",
    "bs4==0.0.2",
    "catalogue==2.0.10",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cloudpathlib==0.21.0",
    "confection==0.1.5",
    "cymem==2.0.11",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "et-xmlfile==2.0.0",
    "faiss-cpu==1.10.0",
    "filelock==3.18.0",
    "frozenlist==1.5.0",
    "fsspec==2025.3.2",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jellyfish==1.2.0",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "joblib==1.4.2",
    "jpype1>=1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.23",
    "langchain-community==0.3.21",
    "langchain-core==0.3.53",
    "langchain-openai==0.3.14",
    "langchain-pinecone==0.2.5",
    "langchain-tests==0.3.19",
    "langchain-text-splitters==0.3.8",
    "langcodes==3.5.0",
    "langsmith==0.3.32",
    "language-data==1.3.0",
    "lark==1.2.2",
    "marisa-trie==1.2.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "murmurhash==1.0.12",
    "mypy-extensions==1.0.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numpy==2.2.4",
    "openai==1.75.0",
    "openpyxl==3.1.5",
    "orjson==3.10.16",
    "outcome==1.3.0.post0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "pinecone==6.0.2",
    "pinecone-plugin-interface==0.0.7",
    "pluggy==1.5.0",
    "preshed==3.0.9",
    "propcache==0.3.1",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.8.1",
    "pygments==2.19.1",
    "pymupdf>=1.25.5",
    "pymupdf4llm>=0.0.22",
    "pysocks==1.7.1",
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "pytest-socket==0.7.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "rake-nltk==1.0.6",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "segtok==1.5.11",
    "selenium==4.31.0",
    "sentence-transformers==4.1.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smart-open==7.1.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.6",
    "spacy==3.8.5",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "sqlalchemy==2.0.40",
    "srsly==2.5.1",
    "sympy==1.13.1",
    "syrupy==4.9.1",
    "tabula-py>=2.10.0",
    "tabulate==0.9.0",
    "tenacity==9.1.2",
    "thinc==8.3.6",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "torch==2.6.0",
    "tqdm==4.67.1",
    "transformers==4.51.3",
    "trio==0.29.0",
    "trio-websocket==0.12.2",
    "typer==0.15.2",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "wasabi==1.1.3",
    "weasel==0.4.1",
    "websocket-client==1.8.0",
    "wrapt==1.17.2",
    "wsproto==1.2.0",
    "yake==0.4.8",
    "yarl==1.20.0",
    "zstandard==0.23.0",
]
